Created Load Manager Class
